+++ scripts/permission_viewer.js
@@ -1,5 +1,60 @@
 class PermissionViewer {
 
+    static init() {
+        JournalSheet.prototype._onShowPlayers = PermissionViewer.prototype._onShowPlayers
+        game.settings.register("permission_viewer", "migrated", {
+            name: game.i18n.localize("PERMISSIONVIEWER.SettingName"),
+            scope: "world",
+            default: 0,
+            type: Number
+        });
+        game.settings.register("permission_viewer", "limitedPrompt", {
+            name: "Limited Permission Dialog",
+            hint: "If checked, limited permission will prompt option dialog.",
+            scope: "world",
+            config: true,
+            default: false,
+            type: Boolean,
+            onChange: value => console.debug(value)
+        });
+
+        // Code Artifact. Keeping in case I want to add Submenu to the settings.
+        // game.settings.registerMenu("permission_viewer", "limitedPrompt", {
+        //     name: "Limited Permission Dialog",
+        //     label: "Permission Viewer Options",
+        //     hint: "Permission Viewer Options",
+        //     type: MySubmenuApplicationClass,
+        //     icon: "fas fa-bars",
+        //     restricted: true
+        // });
+    }
+    static ready() {
+        if (game.settings.get("permission_viewer", "migrated") === 0) {
+            const limnitedJournals = game.journal.entities.filter(j => j.data.permission.default === CONST.ENTITY_PERMISSIONS.LIMITED);
+            if (limnitedJournals > 0) {
+                new Dialog({
+                    "title": game.i18n.localize("PERMISSIONVIEWER.MigrationDialogTitle"),
+                    "content": game.i18n.localize("PERMISSIONVIEWER.MigrationDialogTitle"),
+                    "buttons": {"migrate": {"label": game.i18n.localize("PERMISSIONVIEWER.MigrationDialogLabel1"),
+                                            "callback": () => {
+                                                PermissionViewer.migrateLimitedToObserver();
+                                                game.settings.set("permission_viewer", "migrated", 1);
+                                            }
+                                        },
+                                "no": {"label": game.i18n.localize("PERMISSIONVIEWER.MigrationDialogLabel2"),
+                                        "callback": () => {
+                                                game.settings.set("permission_viewer", "migrated", 1);
+                                            }
+                                        },
+                                },
+                    "default": "migrate"
+                }, {width: 600}).render(true)
+            } else {
+                game.settings.set("permission_viewer", "migrated", 1);
+            }
+        }
+
+    }
     static directoryRendered(obj, html, data) {
         if (!game.user.isGM) return;
         const contextOptions = obj._getEntryContextOptions();
@@ -38,7 +93,6 @@ class PermissionViewer {
                         user_div.addClass("permission-viewer-user")
                     }
                     user_div.css({'background-color': bg_color})
-                    user_div.addClass("pvTooltip");
                     users.push(user_div)
                 }
             }
@@ -71,48 +125,49 @@ class PermissionViewer {
             }
         }
     }
-
     async _onShowPlayers(event) {
         event.preventDefault();
         await this.submit();
         let permissions = this.object.data.permission;
+
         let default_permission = permissions.default || CONST.DOCUMENT_PERMISSION_LEVELS.NONE;
-        if (default_permission >= CONST.DOCUMENT_PERMISSION_LEVELS.LIMITED) {
+        if (default_permission >= CONST.DOCUMENT_PERMISSION_LEVELS.LIMITED && !game.settings.get("permission_viewer","limitedPrompt"))) {
+
             return this.object.show(this._sheetMode, true);
         } else {
             let sharedWith = Object.keys(permissions)
                 .map(id => id == 'default' ? undefined : game.users.get(id))
                 .filter(user => user && permissions[user.id] >= CONST.DOCUMENT_PERMISSION_LEVELS.LIMITED)
-            let buttons = {"show": {"label": "Show to All",
+            let buttons = {"show": {"label": game.i18n.localize("PERMISSIONVIEWER.ShowAll"),
                                     "callback": () => this.object.show(this._sheetMode, true)},
-                           "share": {"label": "Share with All",
-                                     "callback": () => {
-                                         // Need to do a copy of the object, otherwise, the document itself gets changes
-                                         // and the update() doesn't trigger any update on the server.
-                                         permissions = duplicate(permissions);
-                                         permissions["default"] = CONST.DOCUMENT_PERMISSION_LEVELS.OBSERVER;
-                                         // Can't use "permission.default" otherwise it doesn't trigger a journal
-                                         // directory re-render
-                                         this.object.update({permission: permissions})
-                                         this.object.show(this._sheetMode, true);
-                                     }
+                            "share": {"label": game.i18n.localize("PERMISSIONVIEWER.ShareAll"),
+                                        "callback": () => {
+                                            // Need to do a copy of the object, otherwise, the entity itself gets changes
+                                            // and the update() doesn't trigger any update on the server.
+                                            permissions = duplicate(permissions);
+                                            permissions["default"] = CONST.DOCUMENT_PERMISSION_LEVELS.OBSERVER;
+                                            // Can't use "permission.default" otherwise it doesn't trigger a journal
+                                            // directory re-render
+                                            this.object.update({permission: permissions})
+                                            this.object.show(this._sheetMode, true);
+                                        }
                                     }
-                          }
-            let message = "<h3>This Journal Entry is not shared with anyone.</h3>" +
-                "<p>Do you want to share it with all players before showing it,</p>" +
-                "<p>or do you want to show it to all players without sharing it.</p>" +
-                "<p>If you decide to share it, its default permissions will be set as Observer</p>"
+                           }
+            let message = game.i18n.localize("PERMISSIONVIEWER.NotShared1") +
+                game.i18n.localize("PERMISSIONVIEWER.NotShared2") +
+                game.i18n.localize("PERMISSIONVIEWER.NotShared3") +
+                game.i18n.localize("PERMISSIONVIEWER.NotShared4")
             if (sharedWith.length > 0) {
-                message = "<h3>This Journal Entry is shared with the following players.</h3>" +
+                message = game.i18n.localize("PERMISSIONVIEWER.SharedWith1") +
                     "<p><strong>" + sharedWith.map(u => u.name).join(", ") + "</strong></p>" +
-                    "<p>Do you want to share it with all players before showing it,</p>" +
-                    "<p>or do you want to show it to all players without sharing it,</p>" +
-                    "<p>or do you want to show it only to the players that it is already shared with?</p>" +
-                    "<p>If you decide to share it, its default permissions will be set as Observer</p>"
-                buttons["display"] = {"label": "Show to list",
-                                      "callback": () => this.object.show(this._sheetMode, false)}
+                    game.i18n.localize("PERMISSIONVIEWER.NotShared2") +
+                    game.i18n.localize("PERMISSIONVIEWER.NotShared3") +
+                    game.i18n.localize("PERMISSIONVIEWER.SharedWith2") +
+                    game.i18n.localize("PERMISSIONVIEWER.NotShared4")
+                buttons["display"] = {"label": game.i18n.localize("PERMISSIONVIEWER.ShowList"),
+                                        "callback": () => this.object.show(this._sheetMode, false)}
             }
-            new Dialog({"title": "Show Journal Entry to Players",
+            new Dialog({"title": game.i18n.localize("PERMISSIONVIEWER.ShowJournal"),
                         "content": message,
                         "buttons": buttons,
                         "default": "show"
@@ -164,7 +219,6 @@ class PermissionViewer {
                 .map(j => {return {_id: j.id, "permission.default": CONST.DOCUMENT_PERMISSION_LEVELS.OBSERVER}})
         JournalEntry.update(updateData);
     }
-   
     static playerListRendered(){
         console.log("Permission Viewer | Player List");
         let pvUsers = game.users.contents;
@@ -188,6 +242,19 @@ class PermissionViewer {
         }
     }
 }
+
+// class MySubmenuApplicationClass extends FormApplication {
+ 
+//     //Added to squash error in registering Submenu with no type.
+//     // Methods are not currently in use.
+//     // getData() {
+//     //   return game.settings.get('myModuleName', 'myComplexSettingName');
+//     // }
+//     //   _updateObject(event, formData) {
+//     //   const data = expandObject(formData);
+//     //   game.settings.set('myModuleName', 'myComplexSettingName', data);
+//     // }
+// }
 
 Hooks.on('renderJournalDirectory', PermissionViewer.directoryRendered)
 Hooks.on('renderSceneDirectory', PermissionViewer.directoryRendered)